requirements.txt
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2

